// src/lib/kline/types.ts
export interface SymbolInfo {
  ticker: string
  name?: string
  shortName?: string
  exchange?: string
  market?: string
  pricePrecision?: number
  volumePrecision?: number
  priceCurrency?: string
  type?: string
  logo?: string
}

export type Timespan = 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month'

export interface Period {
  timeframe: string;
  multiplier: number;
  timespan: Timespan;
  secs: number;
}